# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A Splitwise-like expense splitting API demonstrating DI, Factory, and Strategy patterns.",
        "title": "Splitwise API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@splitwise.local"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/expenses": {
            "post": {
                "description": "Create an expense with automatic split calculation using EVEN, PERCENTAGE, or EXACT strategy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expenses"
                ],
                "summary": "Create a new expense",
                "parameters": [
                    {
                        "description": "Expense creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/expense.CreateExpenseRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/expense.ExpenseResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/expenses/group/{groupId}": {
            "get": {
                "description": "Get a paginated list of expenses for a group",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expenses"
                ],
                "summary": "List expenses by group",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Group ID",
                        "name": "groupId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Items per page",
                        "name": "per_page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/expense.ExpenseResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/expenses/splits/{splitId}/confirm": {
            "post": {
                "description": "Payer confirms they received the payment",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "splits"
                ],
                "summary": "Confirm split payment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Split ID",
                        "name": "splitId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/expense.SplitResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/expenses/splits/{splitId}/dispute": {
            "post": {
                "description": "Borrower disputes their split with a reason",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "splits"
                ],
                "summary": "Dispute a split",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Split ID",
                        "name": "splitId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Dispute reason",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/expense.DisputeSplitRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/expense.SplitResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/expenses/splits/{splitId}/pay": {
            "post": {
                "description": "Borrower marks their split as paid (waiting for payer confirmation)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "splits"
                ],
                "summary": "Mark split as paid",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Split ID",
                        "name": "splitId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/expense.SplitResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
        "/expenses/{id}": {
            "get": {
                "description": "Get an expense with all its splits",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expenses"
                ],
                "summary": "Get expense by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Expense ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/expense.ExpenseResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an expense (only if no splits are paid/confirmed)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "expenses"
                ],
                "summary": "Delete an expense",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Expense ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "expense.CreateExpenseRequest": {
            "type": "object",
            "required": [
                "amount",
                "description",
                "group_id",
                "participants",
                "split_type"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "description": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                },
                "group_id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "participants": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/expense.SplitParticipant"
                    }
                },
                "split_type": {
                    "type": "string",
                    "enum": [
                        "EVEN",
                        "PERCENTAGE",
                        "EXACT"
                    ]
                }
            }
        },
        "expense.DisputeSplitRequest": {
            "type": "object",
            "required": [
                "reason"
            ],
            "properties": {
                "reason": {
                    "type": "string",
                    "maxLength": 500,
                    "minLength": 1
                }
            }
        },
        "expense.ExpenseResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "group_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "type": "string"
                },
                "payer_id": {
                    "type": "integer"
                },
                "payer_username": {
                    "type": "string"
                },
                "split_type": {
                    "type": "string"
                },
                "splits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/expense.SplitResponse"
                    }
                }
            }
        },
        "expense.SplitParticipant": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "For EXACT split",
                    "type": "number"
                },
                "percentage": {
                    "description": "For PERCENTAGE split",
                    "type": "number"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "expense.SplitResponse": {
            "type": "object",
            "properties": {
                "amount_owed": {
                    "type": "number"
                },
                "borrower_id": {
                    "type": "integer"
                },
                "borrower_username": {
                    "type": "string"
                },
                "dispute_reason": {
                    "type": "string"
                },
                "expense_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "settlement_id": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/expense.SplitStatus"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "expense.SplitStatus": {
            "type": "string",
            "enum": [
                "PENDING",
                "PAID",
                "CONFIRMED",
                "DISPUTED"
            ],
            "x-enum-varnames": [
                "SplitStatusPending",
                "SplitStatusPaid",
                "SplitStatusConfirmed",
                "SplitStatusDisputed"
            ]
        },
...
        "response.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/response.APIError"
                },
                "meta": {
                    "$ref": "#/definitions/response.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.Meta": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your bearer token in the format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}